--- src/log.go
@@ -1,14 +1,0 @@
-package main
-
-import (
-	log "github.com/Sirupsen/logrus"
-)
-
-
-func setLogLevel(level string) {
-	logLevel, err := log.ParseLevel(level)
-	if err != nil {
-		panic(err.Error())
-	}
-	log.SetLevel(logLevel)
-}
